+++ js/web-components/markdown-content.js
@@ -238,6 +238,12 @@ class MarkdownContent extends HTMLElement {
 
         // Set Content
         this.innerHTML = html;
+        if (window.hljs !== undefined) {
+            const codeBlocks = document.querySelectorAll('code[class*="language-"]');
+            for (let code of codeBlocks) {
+                code.classList.add('hljs');
+            }
+        }
         this.dispatchRendered();
     }
 }