+++ CHANGELOG.md
@@ -4,7 +4,7 @@ DataFormsJS uses [Semantic Versioning](https://docs.npmjs.com/about-semantic-ver
 
 Overall the core Framework files and API are expected to remain stable however the version number is expected to increase to much larger numbers in the future due to the changes to smaller scripts and components. This change log includes Framework release history and new website features or major changes.
 
-## Next Release for NPM (Changes on Master Branch)
+## 5.2.0 (November 23, 2020)
 
 * Add React Component `<Markdown>`
   * Based on DataFormsJS Web Component `<markdown-content>`